opt: move dce into its own pass module
