+++ purple-garden-opt/src/ir/mod.rs
@@ -3,9 +3,12 @@
 //! Each pass lives in its own submodule. Orchestration (which passes
 //! run, in what order) lives in [crate::ir] in `src/opt/mod.rs`.
 
+// TODO: come up with Ir::Alloc, Ir::Store, Ir::Load and Ir::AddrOf optimisations
+
 mod branch_cmp;
 mod const_fold;
 mod const_fold_syscalls;
+mod dce;
 mod imm_fold;
 mod indirect_jump;
 mod ret_inline;
@@ -92,47 +95,6 @@ impl<'scratch> Scratch<'scratch> {
     }
 }
 
-/// Remove dead SSA-producing instructions that have no observable effect.
-///
-/// This runs to a fixed point because removing one dead producer can make
-/// earlier producers dead too.
-pub fn dce(fun: &mut ir::Func<'_>, scratch: &mut Scratch<'_>) {
-    loop {
-        let mut changed = false;
-
-        record_uses(fun, scratch);
-
-        for block in &mut fun.blocks {
-            if block.tombstone {
-                continue;
-            }
-
-            for instr in &mut block.instructions {
-                let Some(dst) = ir::Func::def_of(instr) else {
-                    continue;
-                };
-
-                if scratch.use_count(dst) != 0 {
-                    continue;
-                }
-
-                if removable(instr) {
-                    purple_garden_shared::trace!(
-                        "[opt::ir::dce] removed dead definition %v{}",
-                        dst.0
-                    );
-                    *instr = ir::Instr::Noop;
-                    changed = true;
-                }
-            }
-        }
-
-        if !changed {
-            break;
-        }
-    }
-}
-
 /// Recompute whole-function SSA use counts in `scratch`.
 ///
 /// This also calls [`Scratch::ensure`] for definitions with zero uses, so
@@ -158,19 +120,11 @@ pub(super) fn record_uses(fun: &ir::Func<'_>, scratch: &mut Scratch<'_>) {
     }
 }
 
-fn removable(instr: &ir::Instr<'_>) -> bool {
-    match instr {
-        ir::Instr::Call { .. } => false,
-        ir::Instr::Sys { fun, .. } => fun.pure,
-        ir::Instr::Noop => false,
-        _ => true,
-    }
-}
-
 // reexports
 pub use branch_cmp::branch_cmp;
 pub use const_fold::const_fold;
 pub use const_fold_syscalls::const_fold_syscalls;
+pub use dce::dce;
 pub use imm_fold::imm_fold;
 pub use indirect_jump::indirect_jump;
 pub use ret_inline::ret_inline;
